+++ R/transition-states.R
@@ -148,7 +148,7 @@ TransitionStates <- ggproto('TransitionStates', Transition,
           path = transform_path(all_frames, next_state, ease, params$transition_length[i], !!id, enter, exit, match),
           polygon = transform_polygon(all_frames, next_state, ease, params$transition_length[i], !!id, enter, exit, match),
           sf = transform_sf(all_frames, next_state, ease, params$transition_length[i], !!id, enter, exit),
-          stop("Unknown layer type", call. = FALSE)
+          stop(type, ' layers not currently supported by transition_states', call. = FALSE)
         )
       }
     }